+++ lib/runners/processor/eslint.rb
@@ -27,6 +27,10 @@ module Runners
 
     register_config_schema SCHEMA.config
 
+    Config.register_warnings do |config|
+      config.add_warning_for_deprecated_option(analyzer: analyzer_id, old: :dir, new: :target)
+    end
+
     CONSTRAINTS = {
       "eslint" => Gem::Requirement.new(">= 5.0.0", "< 8.0.0").freeze,
     }.freeze
@@ -55,8 +59,6 @@ module Runners
     end
 
     def setup
-      warnings.add_warning_for_deprecated_option(config: config, analyzer: analyzer_id, old: :dir, new: :target)
-
       begin
         install_nodejs_deps constraints: CONSTRAINTS
       rescue UserError => exn