+++ lib/runners/processor/misspell.rb
@@ -20,6 +20,10 @@ module Runners
 
     register_config_schema SCHEMA.config
 
+    Config.register_warnings do |config|
+      config.add_warning_for_deprecated_option(analyzer: analyzer_id, old: :targets, new: :target)
+    end
+
     DEFAULT_TARGET = ".".freeze
 
     def self.config_example
@@ -34,11 +38,6 @@ module Runners
 
     def extract_version_option
       "-v"
-    end
-
-    def setup
-      warnings.add_warning_for_deprecated_option(config: config, analyzer: analyzer_id, old: :targets, new: :target)
-      yield
     end
 
     def analyze(_changes)